Add `:legacy:account` module and move `BaseAccount` there
