+++ feature/widget/shortcut/build.gradle.kts
@@ -5,6 +5,7 @@ plugins {
 dependencies {
     implementation(projects.legacy.ui.legacy)
     implementation(projects.legacy.core)
+    implementation(projects.legacy.account)
 }
 
 android {
