+++ legacy/account/build.gradle.kts
@@ -1,0 +1,11 @@
+plugins {
+    id(ThunderbirdPlugins.Library.android)
+}
+
+android {
+    namespace = "app.k9mail.legacy.account"
+}
+
+dependencies {
+
+}
