+++ legacy/core/src/main/java/com/fsck/k9/Account.kt
@@ -1,5 +1,6 @@
 package com.fsck.k9
 
+import app.k9mail.legacy.account.BaseAccount
 import com.fsck.k9.backend.api.SyncConfig.ExpungePolicy
 import com.fsck.k9.mail.Address
 import com.fsck.k9.mail.ServerSettings
